move reloadSurveyAnsweredRatio and getAnsweredRatio from model to usecase
